# documentation/openapi.json
{
  "openapi": "3.1.0",
  "components": {
    "schemas": {
...
      "EnvironmentVars": {
        "type": "object",
        "description": "An object to hold all the ways environment variables can be passed. Not to be used by itself.",
        "properties": {
          "fromMap": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Map of environment variables to be passed to Deployment.  Ignored if Theia applications are started eagerly.  Empty by default."
          },
          "fromConfigMaps": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of ConfigMaps (by name) containing environment variables to be passed to Deployment as envFrom.configMapRef.  Ignored if Theia applications are started eagerly.  Empty by default."
          },
          "fromSecrets": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of Secrets (by name) containing environment variables to be passed to Deployment as envFrom.secretRef.  Ignored if Theia applications are started eagerly.  Empty by default."
          }
        }
      },
...
      "LaunchRequest": {
        "type": "object",
        "required": ["appId", "user"],
        "description": "A request to launch a new session.",
        "properties": {
          "appId": {
            "type": "string",
            "description": "The App Id of this Theia Cloud instance. Request without a matching Id will be denied."
          },
          "user": {
            "type": "string",
            "description": "The user identification, usually the email address."
          },
          "appDefinition": {
            "type": "string",
            "description": "The app to launch. Needs to be set if a new or ephemeral session should be launched. For an existing workspace the last app definition will be used if none is given."
          },
          "workspaceName": {
            "type": "string",
            "description": "The name of the workspace to mount/create. Needs to be set if an existing workspace should be launched."
          },
          "label": {
            "type": "string",
            "description": "The label of the workspace to mount/create. If no label is given, a default label will be generated."
          },
          "ephemeral": {
            "type": "boolean",
            "description": "If true no workspace will be created for the session."
          },
          "timeout": {
            "type": "integer",
            "format": "int32",
            "description": "Number of minutes to wait for session launch. Default is 3 Minutes."
          },
          "env": {
            "$ref": "#/components/schemas/EnvironmentVars",
            "type": "object",
            "description": "Environment variables"
          }
        }
      },
...
    },
    "securitySchemes": {
      "SecurityScheme": {
        "type": "oauth2",
        "description": "Authentication",
        "flows": {
          "implicit": {}
        }
      }
    }
  },
  "info": {
    "title": "Theia Cloud API",
    "version": "1.2.0"
  },
  "paths": {
    "/service": {
      "post": {
        "summary": "Launch Session",
        "description": "Launches a session and creates a workspace if required. Responds with the URL of the launched session.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LaunchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Not Allowed"
          }
        },
        "tags": ["Root Resource"],
        "security": [
          {
            "SecurityScheme": []
          }
        ]
      }
    },
...
    "/service/{appId}": {
      "get": {
        "summary": "Ping",
        "description": "Replies if the service is available.",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Not Allowed"
          }
        },
        "tags": ["Root Resource"],
        "security": [
          {
            "SecurityScheme": []
          }
        ]
      }
    }
  }
}